feat: shared code and fixed next invoice dates
