+++ backend/package.json
@@ -4,7 +4,7 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "dev": "ts-node-dev --respawn -- src/index.ts",
+    "dev": "ts-node-dev --respawn --require tsconfig-paths/register src/index.ts",
     "dev:test": "TZ=UTC JEST=1 jest --watchAll ",
     "build": "tspc && cp -R ./assets ./dist/ && yarn sentry:sourcemaps",
     "start": "tsc && cp -R ./assets ./dist/ && node --prof ./dist/src/index.js",
@@ -45,6 +45,7 @@
     "json-schema-to-typescript": "^13.1.1",
     "ts-node": "^10.9.2",
     "ts-node-dev": "^2.0.0",
+    "tsconfig-paths": "^4.2.0",
     "typescript": "5.5"
   },
   "imports": {
