+++ backend/src/platform/e-invoices/adapters/superpdp/FAILING.md
@@ -1,70 +1,116 @@
 {
-  "number": "D-2026-2",
-  "issue_date": "2026-04-29",
-  "type_code": 325,
-  "currency_code": "EUR",
-  "vat_category_code": "S",
-  "seller": {
-    "name": "Tricatel",
-    "vat": "",
-    "postal_address": {
-      "street_name": "Avenue de la République",
-      "additional_street_name": "",
-      "city_name": "Chambray-lès-Tours",
-      "postal_zone": "37170",
-      "country": "FR"
-    }
-  },
-  "buyer": {
-    "name": "Test 2",
-    "vat": "",
-    "postal_address": {
-      "street_name": "N/A",
-      "additional_street_name": "",
-      "city_name": "N/A",
-      "postal_zone": "00000",
-      "country": "FR"
-    }
-  },
-  "payment_details": {
-    "payment_means_type_code": "30"
-  },
-  "totals": {
-    "sum_of_invoice_net_amounts": 75,
-    "invoice_total_amount_without_vat": 75,
-    "invoice_total_vat_amount": 0,
-    "invoice_total_amount_with_vat": 75,
-    "amount_due_for_payment": 75,
-    "tax_exclusive_amount": 75,
-    "tax_amount": 0,
-    "tax_inclusive_amount": 75
-  },
-  "vat_break_down": [
-    {
-      "vat_category_taxable_amount": 75,
-      "vat_category_tax_amount": 0,
-      "vat_category_code": "S",
-      "vat_category_rate": 0
-    }
-  ],
-  "lines": [
-    {
-      "line_number": "1",
-      "invoiced_quantity": 1,
-      "invoiced_quantity_unit_code": "HUR",
-      "net_amount": 75,
-      "vat_information": {
-        "invoiced_item_vat_category_code": "S",
-        "invoiced_item_vat_rate": 0
-      },
-      "price_details": {
-        "item_net_price": 75,
-        "base_quantity": 1,
-        "base_quantity_unit_code": "HUR"
-      },
-      "item_information": {
-        "name": "Service par défaut"
-      }
-    }
-  ]
+"process_control": {
+"business_process_type": "M1",
+"specification_identifier": "urn:cen.eu:en16931:2017"
+},
+"number": "D-2026-2",
+"issue_date": "2026-04-29",
+"payment_due_date": "2026-05-12",
+"type_code": 325,
+"currency_code": "EUR",
+"notes": [
+{
+"note": "L'indemnité forfaitaire légale pour frais de recouvrement est de 40 €.",
+"subject_code": "PMT"
+},
+{
+"note": "À défaut de règlement à la date d'échéance, une pénalité de 3 fois le taux légal sera applicable immédiatement.",
+"subject_code": "PMD"
+},
+{
+"note": "Aucun escompte pour paiement anticipé.",
+"subject_code": "AAB"
+}
+],
+"vat_category_code": "S",
+"seller": {
+"name": "Tricatel",
+"vat_identifier": "",
+"postal_address": {
+"address_line1": "Avenue de la République",
+"address_line2": "",
+"city": "Chambray-lès-Tours",
+"post_code": "37170",
+"country_code": "FR"
+},
+"identifiers": [
+{
+"value": "123456789",
+"scheme": "0225"
+}
+],
+"legal_registration_identifier": {
+"value": "123456789",
+"scheme": "0002"
+},
+"electronic_address": {
+"value": "123456789",
+"scheme": "0225"
+}
+},
+"buyer": {
+"name": "AMAZON CREEK MANAGEMENT",
+"vat_identifier": "",
+"postal_address": {
+"address_line1": "160 route du mont",
+"address_line2": "",
+"city": "CHAMONIX-MONT-BLANC",
+"post_code": "74400",
+"country_code": "FR"
+},
+"identifiers": [
+{
+"value": "833840879",
+"scheme": "0225"
+}
+],
+"legal_registration_identifier": {
+"value": "833840879",
+"scheme": "0002"
+},
+"electronic_address": {
+"scheme": "0225",
+"value": "833840879"
+}
+},
+"payment_details": {
+"payment_means_type_code": "30"
+},
+"totals": {
+"sum_invoice_lines_amount": "75",
+"total_without_vat": "75",
+"total_vat_amount": {
+"value": "15"
+},
+"total_with_vat": "90",
+"amount_due_for_payment": "90"
+},
+"vat_break_down": [
+{
+"vat_category_taxable_amount": "75",
+"vat_category_tax_amount": "15",
+"vat_category_code": "S",
+"vat_category_rate": "20"
+}
+],
+"lines": [
+{
+"line_number": "1",
+"invoiced_quantity": 1,
+"invoiced_quantity_unit_code": "HUR",
+"net_amount": 75,
+"vat_information": {
+"invoiced_item_vat_category_code": "S",
+"invoiced_item_vat_rate": 20
+},
+"price_details": {
+"item_net_price": 75,
+"base_quantity": 1,
+"base_quantity_unit_code": "HUR"
+},
+"item_information": {
+"name": "Service par défaut"
+}
+}
+]
 }
