+++ backend/src/services/modules/statistics/services/client-profitability.ts
@@ -6,7 +6,7 @@ import _ from "lodash";
 import Articles, { ArticlesDefinition } from "../../articles/entities/articles";
 import Contacts, { ContactsDefinition } from "../../contacts/entities/contacts";
 import Invoices, { InvoicesDefinition } from "../../invoices/entities/invoices";
-import { getTimezoneOffset } from "../../invoices/utils";
+import { getTimezoneOffset } from "@shared/invoices";
 
 export type TimeRange = {
   label: string;
